New line removed
Yes. A hole commit for a 1 line removal. The line didn't even have any information. why delete??????? Well I hate my code quality not being A+. Did you know that I do not like a letter based grading system. I only know numbers like 10, or 6. About 6 it always felt like 6 was bad and a 7 was average, if you had an 8 or above you did really well. At least in my mind but maybe not in yours!!! Cool now i am going to finish this comment. Hopefully it is not to long!